LC238 (Med) Product of Array Except Self (#10)
Implementation:       
If we could use division, the solution is simple. Could simply iterate
through the array and divide the product by the current number.
Because no division, do the same thing by using a prefix and postfix
array. Every result will be the product of the prefix (numbers before)
and postfix (numbers after).
The prefix array at the current index will be the product of all the
numbers before the current index.
The postfix array at the current index will be the product of all the
numbers after the current index.
Time Complexity: O(2N) = O(N) because we iterate through the array twice
Space Complexity: O(3N) = O(N) because we create three arrays, prefix,
postfix, and results.